+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Bug #561: Fix `ActiveRecordInterface::upsert()` with `$updateProperties = false` (@Tigrov)
 - Bug #550: Relation query should be created by related class, not primary model class (@batyrmastyr)
 - Enh #571: Optimize performance of `ActiveRecord::get()` method (@Tigrov)
+- Enh #576: Add default config for `yiisoft/config` plugin (@Tigrov)
 
 ## 1.0.2 March 11, 2026